+++ src/app/Http/Controllers/QuestionController.php
@@ -15,6 +15,7 @@ class QuestionController extends Controller
      */
     public function index($quizId)
     {
+        $this->authorize("view", Question::class);
         return Question::where(["quiz_id" => $quizId])->get();
     }
 
@@ -27,7 +28,7 @@ class QuestionController extends Controller
     public function store(Request $request, $quizId)
     {
         Quiz::FindOrFail($quizId);
-
+        $this->authorize("create", Question::class);
         $this->validate(
             $request,
             [
@@ -53,6 +54,7 @@ class QuestionController extends Controller
     public function show($quizId, $questionId)
     {
         $question = Question::FindOrFail($questionId);
+        $this->authorize("view", Question::class);
         return $question;
     }
 
@@ -66,6 +68,7 @@ class QuestionController extends Controller
     public function update(Request $request, $quizId, $questionId)
     {
         $question = Question::findOrFail($questionId);
+        $this->authorize("update",Question::class);
         $question->text = $request->text;
 
         $question->save();
@@ -82,6 +85,7 @@ class QuestionController extends Controller
     public function destroy($quizId, $questionId)
     {
         $question = Question::findOrFail($questionId);
+        $this->authorize("delete",Question::class);
         $question->delete();
         return response()->json(["question deleted succesfully"]);
     }
